+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bug fixes:
   - -- Fixed bug where user template dropdown was enabled with no templates
   - -- Fixed bug where switching or copying profiles would not apply profile specific settings
+  - -- Fixed bug where reindexing bars did not sort objectives on hidden buttons
   - -- Removed forgotten print statement when creating new template
 
 ## 3.0-alpha8